add RampdownNoticeFragment
