+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/rampdown/model/StatusTabNotice.kt
@@ -1,0 +1,8 @@
+package de.rki.coronawarnapp.rampdown.model
+
+data class StatusTabNotice(
+    val titleText: String,
+    val subTitleText: String,
+    val longText: String,
+    val faqAnchor: String?
+)
