+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/main/home/HomeFragment.kt
@@ -282,6 +282,12 @@ class HomeFragment : Fragment(R.layout.home_fragment_layout), AutoInject {
                 }
                 openUrl(urlToOpen)
             }
+
+            is HomeFragmentEvents.OpenEolNotice -> {
+                findNavController().navigate(
+                    HomeFragmentDirections.actionMainFragmentToRampdownNoticeFragment(event.statusTabNotice)
+                )
+            }
         }
     }
