+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/main/home/HomeFragmentEvents.kt
@@ -4,6 +4,7 @@ import androidx.annotation.StringRes
 import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.coronatest.type.BaseCoronaTest
 import de.rki.coronawarnapp.coronatest.type.TestIdentifier
+import de.rki.coronawarnapp.rampdown.model.StatusTabNotice
 
 sealed class HomeFragmentEvents {
     data class ShowTracingExplanation(val maxEncounterAgeInDays: Long) : HomeFragmentEvents()
@@ -20,7 +21,7 @@ sealed class HomeFragmentEvents {
     object GoToFamilyTests : HomeFragmentEvents()
     data class GoToTestResultNegativeFragment(val identifier: TestIdentifier) : HomeFragmentEvents()
     data class GoToTestResultKeysSharedFragment(val identifier: TestIdentifier) : HomeFragmentEvents()
-    object OpenEolNotice : HomeFragmentEvents()
+    data class OpenEolNotice(val statusTabNotice: StatusTabNotice) : HomeFragmentEvents()
 
     data class OpenIncompatibleUrl(val scanningSupported: Boolean) : HomeFragmentEvents() {
         @get:StringRes
